收货，无批次小数

+++ app/src/main/java/com/runwise/supply/firstpage/entity/ReceiveBean.java
@@ -16,8 +16,8 @@ import java.util.List;
 public class ReceiveBean {
     private String name;        //货物名称
     private int productId;   //商品id
-    private int count;          //商品数量
-    private int productUomQty;          //商品库存数量
+    private double count;          //商品数量
+    private double productUomQty;          //商品库存数量
     private boolean isTwoUnit;  //是否为双单位
     private String unit;        //是双单位的话，单位是什么
     private double twoUnitValue;//有双单位情况下，用户输入的值
@@ -31,8 +31,8 @@ public class ReceiveBean {
             JSONObject jsonObject = new JSONObject(source);
             name = jsonObject.optString("name");
             productId = jsonObject.optInt("productId");
-            count = jsonObject.optInt("count");
-            productUomQty = jsonObject.optInt("productUomQty");
+            count = jsonObject.optDouble("count");
+            productUomQty = jsonObject.optDouble("productUomQty");
             isTwoUnit = jsonObject.optBoolean("isTwoUnit");
             twoUnitValue = jsonObject.optDouble("twoUnitValue");
             unit = jsonObject.optString("unit");
@@ -90,11 +90,11 @@ public class ReceiveBean {
 
     private String defaultCode;
 
-    public int getProductUomQty() {
+    public double getProductUomQty() {
         return productUomQty;
     }
 
-    public void setProductUomQty(int productUomQty) {
+    public void setProductUomQty(double productUomQty) {
         this.productUomQty = productUomQty;
     }
 
@@ -125,11 +125,11 @@ public class ReceiveBean {
         this.productId = productId;
     }
 
-    public int getCount() {
+    public double getCount() {
         return count;
     }
 
-    public void setCount(int count) {
+    public void setCount(double count) {
         this.count = count;
     }
